helpers/dom -> helpers, exterminate old layout
