+++ src/components/DropAndCrop/Cropper.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
 import {BehaviorSubject} from 'rx'
-import {reactComponent} from 'helpers/dom'
+import {reactComponent} from 'helpers'
 
 // stupid react component requires ref (and thus a class)
 // to get dataurl from canvas
