+++ src/root/Admin/index.js
@@ -9,8 +9,8 @@ import Header from 'components/Header'
 import TabBar from 'components/TabBar'
 
 import {nestedComponent, mergeOrFlatMapLatest} from 'util'
-import {icon} from 'helpers/dom'
-import {mobileLayout, desktopLayout} from 'helpers/layout'
+import {icon} from 'helpers'
+// import {mobileLayout, desktopLayout} from 'helpers/layout'
 
 import ComingSoon from 'components/ComingSoon'
 
@@ -77,55 +77,3 @@ export default sources => {
     route$,
   }
 }
-
-// const NavContent = sources => ({
-//   DOM: Observable.just(div({},'nav content')),
-// })
-
-// const _DOM = ({
-//   pageDOM, appBarDOM, tabBarDOM, navContentDOM, isMobile, isOpen,
-// }) =>
-//   (isMobile ? mobileLayout : desktopLayout)({
-//     bar: appBarDOM,
-//     tabs: tabBarDOM,
-//     side: navContentDOM,
-//     main: pageDOM,
-//     isOpen,
-//   })
-
-// export default sources => {
-//   const appBar = AppBar(sources) // will need to pass auth
-//   const tabBar = TabBar({...sources, tabs: Observable.just(_tabs)})
-//   const navContent = NavContent(sources)
-
-//   const page$ = nestedComponent(sources.router.define(_routes),sources)
-
-//   const children = [appBar,tabBar,navContent,page$]
-
-//   const closeSideNav$ = sources.DOM.select('.close-sideNav').events('click')
-
-//   const sidenavOpen$ = appBar.navButton$.map(true)
-//     .merge(closeSideNav$.map(false))
-//     .startWith(false)
-
-//   const DOM = combineLatestObj({
-//     pageDOM$: page$.pluck('DOM'),
-//     appBarDOM$: appBar.DOM,
-//     tabBarDOM$: tabBar.DOM,
-//     navContentDOM$: navContent.DOM,
-//     isMobile$: sources.isMobile$,
-//     isOpen: sidenavOpen$,
-//   }).map(_DOM)
-
-//   const route$ = Observable.merge(
-//     mergeOrFlatMapLatest('route$',...children),
-//     sources.redirectLogout$
-//   )
-
-//   return {
-//     DOM,
-//     auth$: mergeOrFlatMapLatest('auth$',...children),
-//     queue$: mergeOrFlatMapLatest('queue$',...children),
-//     route$,
-//   }
-// }
