+++ src/util.js
@@ -19,3 +19,41 @@ export const mergeOrFlatMapLatest = (prop, ...sourceArray) =>
         src[prop] || Observable.empty()
     )
   )
+
+// app-wide material styles
+export const material = {
+  primaryColor: '#666',
+  primaryFontColor: 'rgba(255, 255, 255, 0.9)',
+  primaryFontColorDisabled: 'rgba(0, 0, 0, 0.45)',
+  primaryLightWaves: false,
+  secondaryColor: '#009688',
+  secondaryFontColor: 'rgba(255, 255, 255, 0.9)',
+  secondaryFontColorDisabled: 'rgba(255, 255, 255, 0.6)',
+  secondaryLightWaves: true,
+  errorColor: '#C00',
+  successColor: '#090',
+  typographyColor: '#212121',
+
+  sidenav: {
+    width: '280px',
+    left: '-290px',
+    transition: 'left .3s ease-out',
+    delayed: {
+      left: '0',
+    },
+    remove: {
+      left: '-290px',
+    },
+  },
+
+  fadeInOut: {
+    opacity: '0',
+    transition: 'opacity .3s',
+    delayed: {
+      opacity: '1',
+    },
+    remove: {
+      opacity: '0',
+    },
+  },
+}
